% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSVAR.R
\name{plot_network}
\alias{plot_network}
\title{plot sparse component for use igraph and network layout}
\usage{
plot_network(mat, threshold = 0.1)
}
\arguments{
\item{mat}{a p by p matrix, indicating the sparse component}

\item{threshold}{the threshold for presenting the edges in the network}
}
\value{
A network plot for the sparse component
}
\description{
Plot a network to illustrate the estimated sparse component
}
\examples{
set.seed(1)
est_mats <- matrix(rnorm(400, 0, 1), 20, 20)
plot_network(est_mats, threshold = 0.1)
}
